requirements.txt
streamlit
langchain-community
langchain-google-genai
google-genai
python-dotenv
numpy
faiss-cpu # or faiss-gpu if deploying to a GPU environment